Refactor hmcts-annotation-ui-lib to em-viewer-ui
