+++ src/app/routing/pages/decisions/check-decision/check-decision.component.ts
@@ -3,10 +3,10 @@ import { DecisionService } from '../../../../domain/services/decision.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { FormsService } from '../../../../shared/services/forms.service';
 import { FormGroup } from '@angular/forms';
-import { NpaService } from '../../../../shared/components/hmcts-annotation-ui-lib/data/npa.service';
-import { IDocumentTask } from '../../../../shared/components/hmcts-annotation-ui-lib/data/document-task.model';
-import { ApiHttpService } from '../../../../shared/components/hmcts-annotation-ui-lib/data/api-http.service';
-import { AnnotationStoreService } from '../../../../shared/components/hmcts-annotation-ui-lib/data/annotation-store.service';
+import { NpaService } from '../../../../shared/components/hmcts-em-viewer-ui/data/npa.service';
+import { IDocumentTask } from '../../../../shared/components/hmcts-em-viewer-ui/data/document-task.model';
+import { ApiHttpService } from '../../../../shared/components/hmcts-em-viewer-ui/data/api-http.service';
+import { AnnotationStoreService } from '../../../../shared/components/hmcts-em-viewer-ui/data/annotation-store.service';
 
 @Component({
     selector: 'app-check-decision',
