+++ src/app/shared/components/hmcts-em-viewer-ui/data/pdf-adapter.ts
