+++ src/app/shared/components/hmcts-em-viewer-ui/data/utils.ts
