+++ src/app/shared/components/hmcts-em-viewer-ui/document-viewer/document-viewer.service.ts
