+++ src/app/shared/components/hmcts-em-viewer-ui/viewers/annotation-pdf-viewer/annotation-pdf-viewer.component.scss
