+++ src/app/shared/components/hmcts-em-viewer-ui/viewers/annotation-pdf-viewer/comments/comment-item/comment-item.component.scss
