+++ src/app/shared/components/hmcts-em-viewer-ui/viewers/annotation-pdf-viewer/contextual-toolbar/contextual-toolbar.component.scss
