+++ src/app/shared/components/hmcts-em-viewer-ui/viewers/annotation-pdf-viewer/contextual-toolbar/contextual-toolbar.component.ts
@@ -1,8 +1,8 @@
 import {Component, OnInit, ChangeDetectorRef, OnDestroy } from '@angular/core';
 import { Subscription } from 'rxjs';
-import { AnnotationStoreService } from '../../data/annotation-store.service';
-import { Annotation, Comment } from '../../data/annotation-set.model';
-import { PdfService } from '../../data/pdf.service';
+import { Annotation, Comment } from '../../../data/annotation-set.model';
+import { AnnotationStoreService } from '../../../data/annotation-store.service';
+import { PdfService } from '../../../data/pdf.service';
 
 @Component({
   selector: 'app-contextual-toolbar',
