+++ src/app/shared/components/hmcts-em-viewer-ui/viewers/image-viewer/image-viewer.component.spec.ts
@@ -6,7 +6,7 @@ class MockRenderer {
 
 }
 
-fdescribe('ImageViewerComponent', () => {
+describe('ImageViewerComponent', () => {
     let component: ImageViewerComponent;
     let fixture: ComponentFixture<ImageViewerComponent>;
     const mockRenderer = new MockRenderer();
