+++ src/app/shared/components/hmcts-em-viewer-ui/viewers/image-viewer/image-viewer.component.ts
