+++ src/app/shared/components/hmcts-em-viewer-ui/viewers/viewer-factory.service.ts
@@ -1,10 +1,10 @@
 import {ComponentFactoryResolver, ComponentRef, Injectable, ViewContainerRef} from '@angular/core';
-import {AnnotationPdfViewerComponent} from '../../hmcts-annotation-ui-lib/components/annotation-pdf-viewer/annotation-pdf-viewer.component';
-import {AnnotationStoreService} from '../../hmcts-annotation-ui-lib/data/annotation-store.service';
-import {IAnnotationSet} from '../../hmcts-annotation-ui-lib/data/annotation-set.model';
-import { ImageViewerComponent } from '../../hmcts-annotation-ui-lib/components/image-viewer/image-viewer.component';
-import { UnsupportedViewerComponent } from '../../hmcts-annotation-ui-lib/components/unsupported-viewer/unsupported-viewer.component';
 import { UrlFixerService } from '../data/url-fixer.service';
+import { AnnotationStoreService } from '../data/annotation-store.service';
+import { ImageViewerComponent } from './image-viewer/image-viewer.component';
+import { UnsupportedViewerComponent } from './unsupported-viewer/unsupported-viewer.component';
+import { IAnnotationSet } from '../data/annotation-set.model';
+import { AnnotationPdfViewerComponent } from './annotation-pdf-viewer/annotation-pdf-viewer.component';
 
 @Injectable()
 export class ViewerFactoryService {
